Replace code to 'InterfaceSettings' component
